Multithreading and Bug Fixing
Added the ability to multi-thread disassembled dataset generation.

Fixed bug in disassembly dataset generation where empty files were being saved after the disassembly process failed.

Implemented the ability to force the transformer to train on the CPU.

Tensorflow is no longer imported when running the disassembler, reducing the overhead for the operation.

Signed-off-by: Anna <wheelercs@protonmail.com>