Single Op Simplify and Performance Improvement
Added and modified options for genericizing disassembled text before writing it to a file. As this process stands, a test dataset saw an 80% reduction in the number of unique instructions. There is still more work to be done on this, especially with handling more commands, pointer types, and some potential expansion to more accurately represent the command.

Additionally, a major performance issue with the capstone framework was rectified. Capstone no longer disassembles the entire executable and only $n$ of the operations are written to a file, but now capstone is limited by $n$ internally, ensuring only part of the file is disassembled and boosting overall disassembly performance.

Signed-off-by: Anna <wheelercs@protonmail.com>