More Efficient Genericization
Added more commands to be recognized by the genericization process. This has led to an approx. 75% reduction over the previous version in instructions created during the generation process.

Code was also cleaned up, with arrays and dictionaries now being used to identify commands and replacement patterns. This should hopefully make it easier to modify and add to the genericizer.

Signed-off-by: Anna <wheelercs@protonmail.com>